Analysis of Sales Report of a Clothes Manufacturing Outlet
DESCRIPTION

Background and Objective:
A high-end fashion retail store is looking to expand its products. It wants to understand the market and find the current trends in the industry. It has a database of all products with attributes, such as style, material, season, and the sales of the products over a period of two months.

Domain: Retail

Dataset Description:
There are two files provided, and the detailed description of each is given below:

 

Attribute 	Description
Dress_ID 	A unique identifier for each dress
Style	Style of dress can belong to one of 12 styles, including casuals, novelty, etc.
Price 	Price category of the dress (low, average, medium, high, and very high)
Rating	A number between 0 and 5, specifying the rating of the dress
Size	Size of the dress (small, medium, large, XL, and free)
Season	Season category of the dress, i.e., summer, spring, etc
Neckline	Type of neckline, for example, V-neck, collar, etc
Sleeve length	Length of the sleeve—full, three-quarters, etc
Waistline	The waistline of the dress
Material 	The material of the dress, for example, silk, cotton, etc
Fabric type 	Fabric type of dress
Decoration	The decoration of the dress, like ruffles, embroidery, etc
Pattern Type	Pattern type of the dress—dot, animal print, etc
Recommendation	A binary value suggesting a recommendation (1) or not (0)
The remaining columns depict the sales for each dress on a particular date.
Date ranges from 29/8/2013 to 12/10/2013, and the sales are registered for alternative days.

Analysis Tasks:
The goals of this project are:

To automate the process of recommendations, the store needs to analyze the given attributes of the product, like the style, season, etc., and come up with a model to predict the recommendation of products (in binary output – 0 or 1) accordingly.
In order to stock the inventory, the store wants to analyze the sales data and predict the trend of total sales for each dress for an extended period of three more alternative days.
To decide the pricing for various upcoming clothes, they wish to find how the style, season, and material affect the sales of a dress and if the style of the dress is more influential than its price.
Also, to increase sales, the management wants to analyze the attributes of dresses and find which are the leading factors affecting the sale of a dress.
To regularize the rating procedure and find its efficiency, the store wants to find if the rating of the dress affects the total sales